# docs/swagger.yaml
basePath: /app/classify/api/
definitions:
  api.Device:
    properties:
      area_id:
        description: 区域
        type: string
      created:
        type: string
      desc:
        type: string
      disabled:
        type: boolean
      gateway_id:
        type: string
      group_id:
        description: 分组
        type: string
      id:
        description: ClientID
        type: string
      name:
        type: string
      parameters:
        additionalProperties:
          type: number
        description: 模型参数，用于报警检查
        type: object
      product_id:
        type: string
      type_id:
        description: 类型
        type: string
    type: object
  api.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
  api.ReplyData-api_Device:
    properties:
      data:
        $ref: '#/definitions/api.Device'
      error:
        type: string
    type: object
...
  api.ReplyList-api_Device:
    properties:
      data:
        items:
          $ref: '#/definitions/api.Device'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: API文档
  title: 设备分类数据接口文档
  version: 1.0 版本
paths:
  /device/{id}:
    get:
      consumes:
      - application/json
      description: 获取设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-api_Device'
      summary: 获取设备
      tags:
      - device
    post:
      consumes:
      - application/json
      description: 修改设备
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      - description: 设备信息
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/api.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-api_Device'
      summary: 修改设备
      tags:
      - device
...
  /device/list:
    get:
      consumes:
      - application/json
      description: 查询设备
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-api_Device'
      summary: 查询设备
      tags:
      - device
  /device/search:
    post:
      consumes:
      - application/json
      description: 查询设备
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-api_Device'
      summary: 查询设备
      tags:
      - device
...
swagger: "2.0"